+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Changelog
 
-## 0.0.6
+## 0.0.7
 
 - Added subjectPrerequisite model
 - Added subjectsPrerequisite routes
 - Added rest api endpoint at server.js
+
+## 0.0.6
+
+- Fixed the error handling in campusCollege.js
+- Fixed the error handling in cashier.js
+- Fixed the error handling in college.js
+- Fixed the error handling in collegeCourses.js
+- Fixed the error handling in course.js
+- Fixed the error handling in document.js
+- Fixed the error handling in mainCampus.js
 
 ## 0.0.5